Move validation tests to test_search
